+++ test/libsolidity/smtCheckerTests/functions/functions_trivial_condition_for_only_call.sol
@@ -1,0 +1,7 @@
+pragma experimental SMTChecker;
+
+contract C
+{
+	function f(bool x) public pure { for (;x;) {} }
+	function g() public pure { f(true); }
+}