+++ test/libsolidity/smtCheckerTests/functions/functions_trivial_condition_require.sol
@@ -1,0 +1,8 @@
+pragma experimental SMTChecker;
+
+contract C
+{
+	function f(bool x) public pure { x = true; require(x); }
+}
+// ----
+// Warning: (98-99): Condition is always true.